add moviecontroller
